requirements.txt
Django==2.0.4
django-filter==1.1.0
django-json-response==1.1.5
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Markdown==2.6.11
Pillow==5.1.0
PyJWT==1.6.1
pytz==2018.4
